Move search results table related source files into a separate directory.
