+++ components/webui/imports/ui/SearchView/SearchResults/SearchResultsTable/SearchResultsHeader.scss
