+++ components/webui/imports/ui/SearchView/SearchResults/SearchResultsTable/SearchResultsTable.scss
